refactor : 정답률 불러오기 해결, 문제 생성 DTO 변결
