+++ cs25-service/src/main/java/com/example/cs25service/domain/admin/controller/QuizAdminController.java
@@ -34,6 +34,7 @@ public class QuizAdminController {
     public ApiResponse<String> uploadQuizByJsonFile(
         @RequestParam("file") MultipartFile file,
         @RequestParam("categoryType") String categoryType,
+        @RequestParam("formatType") QuizFormatType formatType,
         @AuthenticationPrincipal AuthUser authUser
     ) {
         if (file.isEmpty()) {
@@ -45,7 +46,7 @@ public class QuizAdminController {
             return new ApiResponse<>(400, "JSON 파일만 업로드 가능합니다.");
         }
 
-        quizAdminService.uploadQuizJson(file, categoryType);
+        quizAdminService.uploadQuizJson(file, categoryType, formatType);
         return new ApiResponse<>(200, "문제 등록 성공");
     }
